docs(examples): cleanup and update jit-parcel
