+++ packages/examples/jit-parcel/.gitignore
@@ -1,4 +1,2 @@
 dist
 .cache
-package-lock.json
-debug.log
