+++ packages/examples/jit-parcel/.npmrc
@@ -1,0 +1,1 @@
+package-lock=false
