+++ packages/examples/jit-parcel/index.html
@@ -5,12 +5,11 @@
   <meta charset="utf-8">
   <title>Aurelia</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.3.3/semantic.min.css" />
 </head>
 
 <body>
-  <app class="ui container"></app>
-  <script src="startup.ts"></script>
+  <app></app>
+  <script src="src/startup.ts"></script>
 </body>
 
 </html>
