+++ packages/examples/jit-parcel/package.json
@@ -2,32 +2,19 @@
   "name": "jit-parcel",
   "version": "0.1.1",
   "scripts": {
-    "build": "rimraf dist && parcel build index.html --no-cache --no-minify --experimental-scope-hoisting --no-source-maps",
-    "serve": "rimraf dist && parcel index.html --no-cache --no-source-maps",
-    "e2e": "mocha ./e2e/**/*.spec.ts --require ts-node/register -R spec -u bdd --recursive -c --check-leaks -t 10000 --watch-extensions ts",
-    "e2e:watch": "npm run e2e -- -w"
+    "watch": "parcel index.html -p 9000 --open",
+    "build": "rimraf dist && parcel build index.html",
+    "serve": "http-server -c-1 -p 9000 dist"
   },
   "dependencies": {
-    "@aurelia/debug": "file:../../debug/aurelia-debug-0.1.1.tgz",
-    "@aurelia/jit": "file:../../jit/aurelia-jit-0.1.1.tgz",
-    "@aurelia/kernel": "file:../../kernel/aurelia-kernel-0.1.1.tgz",
-    "@aurelia/runtime": "file:../../runtime/aurelia-runtime-0.1.1.tgz"
+    "@aurelia/jit": "dev"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.4",
-    "@types/mocha": "^5.2.5",
     "@types/node": "^8.10.22",
-    "@types/selenium-webdriver": "^3.0.10",
-    "chai": "^4.1.2",
-    "chromedriver": "^2.41.0",
-    "cross-env": "^5.2.0",
-    "mocha": "^5.2.0",
-    "parcel-bundler": "^1.9.7",
-    "path": "^0.12.7",
-    "rimraf": "^2.6.2",
-    "selenium-webdriver": "^4.0.0-alpha.1",
-    "ts-node": "^7.0.1",
-    "tsconfig-paths": "^3.5.0",
-    "typescript": "^3.0.1"
+    "http-server": "latest",
+    "parcel-bundler": "git+https://github.com/parcel-bundler/parcel.git#master",
+    "parcel-plugin-typescript": "latest",
+    "rimraf": "latest",
+    "typescript": "latest"
   }
 }
