+++ packages/examples/jit-parcel/src/app.html
@@ -1,0 +1,3 @@
+<template>
+  <div>${message}</div>
+</template>
