+++ packages/examples/jit-parcel/src/app.ts
@@ -1,0 +1,15 @@
+import { customElement } from '@aurelia/runtime';
+import view from './app.html';
+
+@customElement({
+  name: 'app',
+  templateOrNode: view,
+  build: {
+    required: true,
+    compiler: 'default'
+  },
+  instructions: []
+})
+export class App {
+  message = 'Hello World!';
+}
