+++ packages/examples/jit-parcel/src/startup.ts
@@ -1,9 +1,8 @@
-import { DebugConfiguration } from '@aurelia/debug';
 import { BasicConfiguration } from '@aurelia/jit';
 import { Aurelia } from '@aurelia/runtime';
 import { App } from './app';
 
 window['au'] = new Aurelia()
-  .register(BasicConfiguration, DebugConfiguration)
+  .register(BasicConfiguration)
   .app({ host: document.querySelector('app'), component: new App() })
   .start();
