+++ packages/examples/jit-parcel/tsconfig.json
@@ -1,8 +1,13 @@
 {
-  "extends": "../../tsconfig-build.json",
   "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "dist"
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true,
+    "lib": ["esnext", "dom"],
+    "module": "esnext",
+    "moduleResolution": "node",
+    "outDir": "dist",
+    "rootDir": "src",
+    "target": "es2018"
   },
-  "include": ["."]
+  "include": ["src"]
 }
